lib/requirements.txt
pytest
pytest-docker
neo4j
pyyaml
colored
agnos
openai
python-dotenv
networkx
pandas
yfinance
psutil
fastapi
uvicorn
python-multipart
requests
